feat: sms preview in editor
